poplar_recovery_builder.sh: additional verification
Add some additional checks throughout, mostly to allow us to make
assumptions that certain valies will not be zero.  This is preparing
for the next patch, which fixes some calls to "expr" so we don't
assume it'll return non-zero status.  (If the the value of the
expression computed by "expr" is 0, its exit status is 1.)

Signed-off-by: Alex Elder <elder@linaro.org>