poplar_recovery_builder.sh: fix expr calls
Make sure our calls to "expr" never produce non-zero status under
otherwise normal circumstances.  This is to work around the fact
that "expr" returns 1 if the expression it's given produces the
value 0.  With this patch in place, we can turn on the "errexit"
option to bash, which exits in the event of *any* error.

Signed-off-by: Alex Elder <elder@linaro.org>