poplar_recovery_builder.sh: write an image
Rather than create a directory full of files to copy to a USB drive,
create another image to be copied (with dd) directly on top of the
USB drive.  That means creating an empty image file, setting it up
with a loop device, partitioning it and making a FAT32 file system,
and then copying our installer files onto that instead of into the
USB directory.

Some help functions are created, because now we need superuser
access to write files into what was previously an unprivileged USB
directory.

Signed-off-by: Alex Elder <elder@linaro.org>