poplar_recovery_builder.sh: use USB image
The code was always using the disk image rather than setting up the
USB image as it should.

Also change the USB device number used for the install script to 0:1.

Have the cleanup function be called on normal exits.  And get rid of
the trap_cleanup() function; incorporate all cleanup in one function.

Signed-off-by: Alex Elder <elder@linaro.org>