poplar_recovery_builder.sh: skip /a and /b
It takes a LONG time to install the file systems.  We aren't
actually using /a and /b, so just skip defining them in our
configuration.

In addition, don't mark the partition on the USB stick as bootable.
It doesn't seem to boot anyway, and we're handling the installation
manually for now (until we figure out why it's not automated like it
was before...)

Add "mem=1G" and "earlycon" to the appended kernel arguments.

And clean up a little white space.

Signed-off-by: Alex Elder <elder@linaro.org>