poplar_recovery_builder.sh: tweak install script commands
The install script was getting the size of each file before it was
loaded.  This was an artifact from an earlier attempt to use the
"gzwrite" command for writing (which had a few problems so I
abandonded it).  Since we're reading the entire file we have no need
to get its size.  So stop needlessly doing the "fatsize" command.

On the other hand, we do know the size of the uncompressed output
file at the time we're generating the installer script.  So we can
add that piece of information to the installer's "mmc_write"
command.

Signed-off-by: Alex Elder <elder@linaro.org>