poplar_recovery_builder.sh: add Android imaging
Arrange to allow an Android image to be created rather than using a
Linux distro.  The Android image is selected by passing "android"
instead of the name of a root file system file.  In that case, we
use various Android image files to populate Android partitions
rather than just gathering the kernel and DTB, etc.

The loader partition is populated regardless of whether we're using
Android or "just" Linux.

  #    Start     Size     Type FS Type Description
  -    -----     ----     ---- ------- -----------
  *        0        1      MBR
  1        1     8191  primary    none loader
  2     8192    81920  primary    none android_boot
  3    90112  2097151  primary    ext4 android_system
  4  2187263 13082625 extended
  *  2187263        1      EBR
  5  2187264  2097151  logical    ext4 android_cache
  *  4284415        1      EBR
  6  4284416 10985472  logical    ext4 android_user_data
  Total EMMC size is 15269888 512-byte sectors

Signed-off-by: Alex Elder <elder@linaro.org>