build_instructions.md: skip the "unmount" step
There's no real need to unmount partitions on the flash drive that's
about to be overwritten.  Linux should properly deal with the drive
contents suddenly getting blown away.  The command that was there
was OK, but if there was nothing mounted it gave confusing errors.
So simplify things by just getting rid of that instruction.

Signed-off-by: Alex Elder <elder@linaro.org>