build_instructions.md: add back a *critical* piece of information
In this commit:
  33f17a1 build_instructions.txt: document workaround
I changed the text to document how one manually initiates the
installation procedure.  Previously it was assumed that this would
occur automatically, but that proved unreliable.

In making these updates I inadvertently deleted the section that
explained the need to press the USB_BOOT button in order to boot
from the memory stick.  THIS WAS A TERRIBLE MISTAKE.

Correct this error.

Signed-off-by: Alex Elder <elder@linaro.org>