poplar_recovery_builder.sh: support other root archives
Andreas Färber mentioned that he had trouble using this script when
supplying a different root file system archive (an OpenSUSE one)
than the Linaro one suggested.  He pointed out two key differences
in the archive contents:  the Linaro archive was compressed with
gzip, while the SUSE one used bzip; and the Linaro archive enclosed
all of its contents in a directory "binary", while the SUSE one does
not.

This patch makes two changes to allow both types of archive to be
supported.  First, a new function unpack() is used to decompress a
file, using its file name extension as a guide to what program
should be used to produce uncompressed output.  Second, the
"--transform" feature of tar(1) is used to cause the leading
"binary" directory name in a Linaro archive to be stripped if it is
present.

To support the latter change, we now add "-C ${MOUNT}" to the tar
extraction command, so the extraction takes place directly inside
the mount point.  There is therefore no reason to use "binary" as
the mount point name, so that is changed to the more instructive
"mount".  A few spots are changed to reflect that we're dealing
with an archive rather than some sort of image file.

With this change, it is possible to produce a working recovery stick
supplying either an Linaro or an OpenSUSE root file system archive.

Suggested-by: Andreas Färber <afaerber@suse.de>
Signed-off-by: Alex Elder <elder@linaro.org>