poplar_recovery_builder.sh: description, not mount point
Generalize the third argument provided to partition_define() to be a
description of the partition rather than explicitly being a mount point.
Mandate that a description be provided to partition_init(), and provide
one for the loader partition.  This is looking ahead to Android, where
we'll describe partitions rather than their mount points.  For the time
being, it's perfectly fine to use this as a mount point--we just won't
call it that (except in fstab_add()), and will no longer insist it begins
with a '/' character.

Signed-off-by: Alex Elder <elder@linaro.org>