poplar_recovery_builder.sh: hierarchical installer
Add the ability to define sub-scripts that allow sets of
installation commands to be grouped logically.  The top-level
script loads a script into a new memory area, then calls the
U-Boot "source" command to to execute it.

Use this capability to encapsulate setting up partition metadata
(MBR and EBRs).  Also use it to encapsulate commands to set up each
partition.

Signed-off-by: Alex Elder <elder@linaro.org>