poplar_recovery_builder.sh: no sudo in cleanup
This patch reverts this change:
    be17e72 poplar_recovery_builder: Fix cleanup

Andreas argued that the rootfs tarball was extracted using sudo, and
therefore the cleanup needed to be done with sudo as well.

The reason sudo was required to do the extraction was because it was
extracting it into the mounted file system.  But by the time the
mount point is removed in cleanup(), that file system has been
unmounted.  And the mount directory that remains was created without
superuser privilege.

Requiring sudo in the cleanup function has the unfortunate property
that an early abort of the script prompts for superuser access in
the cleanup function even when it's not required.  More than one
person has expressed concern about this when it has occurred.

So revert the change.

Signed-off-by: Alex Elder <elder@linaro.org>