poplar_recovery_builder.sh: update Android partition sizes
When support for Android images were added, the commit header
described the intended sizes of partitions to be used:
  e5271fe poplar_recovery_builder.sh: add Android imaging

Despite that, the sizes actually assigned were different.  In
particular, the system and cache partitions were one 512-byte sector
larger than intended, and the user data partition was much smaller
than it could be (leaving unused a large part of the eMMC media
available).  The partition sizes that were implemented came from the
definitions found in the Poplar "BoardConfig.mk" file in the Android
tree.  But again, these did not match the sizes that were intended
(which were described in the commit message).

It was not possible to simply reduce these partition sizes.  Unless
"BoardConfig.mk" was modified to reduce the defined image sizes
first, the Android images built would be too large for the
partitions created here.

As of this commit in the "devices/hisilicon/poplar" Android sub-tree:
  xxxxxxx poplar: fix partition sizes
the Android code has been updated so the images that get built will
match (and won't exceed) the sizes defined by this patch.

Signed-off-by: Alex Elder <elder@linaro.org>