poplar_recovery_builder.sh: switch to new Ethernet version
Swith the main recovery builder script over so it uses Ethernet for
rather than USB memory stick for transferring data.  The USB stick
is only required for de-brick, and in that case, only "fastboot.bin"
is needed.  Once de-bricked, the rest of the system can be recovered
using Ethernet.

There are some other changes to create another partition to hold the
Android boot image, which will allow us to boot Android using UEFI.
But for now those changes cannot be considered fully tested.

Signed-off-by: Alex Elder <elder@linaro.org>